rename : Configuration -> Config로 클래스, 디렉토리 명 통일
