+++ src/main/java/com/grablunchtogether/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.grablunchtogether.configuration;
+package com.grablunchtogether.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -16,7 +16,7 @@ import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 @Configuration
 @EnableSwagger2
-public class SwaggerConfiguration implements WebMvcConfigurer {
+public class SwaggerConfig implements WebMvcConfigurer {
     @Bean
     public Docket api() {
