+++ src/main/java/com/grablunchtogether/service/UserService.java
@@ -1,6 +1,6 @@
 package com.grablunchtogether.service;
 
-import com.grablunchtogether.configuration.JwtTokenProvider;
+import com.grablunchtogether.config.JwtTokenProvider;
 import com.grablunchtogether.domain.User;
 import com.grablunchtogether.dto.geocode.GeocodeDto;
 import com.grablunchtogether.dto.token.TokenDto;
